Add Basic Biomes

+++ Assets/Scripts/Terrain/TileData.cs
@@ -25,7 +25,7 @@ public class TileData
 
     public TileType type;
     public float elevationValue = 0;
-    public int moistureValue = 0;
+    public float moistureValue = 0;
     public float temperatureValue = 0f;
 
     public TileData(int xCoordinate, int zCoordinate)
